Rename monkey patch
