+++ lib/action_view/component.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
 require "action_view/component/monkey_patch"
-require "action_view/component/am_monkey_patch"
+require "action_view/component/active_model_conversion_monkey_patch"
 require "action_view/component/base"
 require "action_view/component/railtie"
